Move each class to its own class
